update microscope configuration schema
